+++ amadeus-integration-test/TravelRestrictionsTest.cs
@@ -13,7 +13,7 @@ namespace amadeus_integration_test
         public void Get_Restrictions()
         {
             var amadeus = GetAmadeusBuild();
-            DiseaseAreaReport response = amadeus.dutyOfCare.travelRestrictions.get(Params.with("countryCode", "US"));
+            DiseaseAreaReport response = amadeus.dutyOfCare.diseases.covid19AreaReport.get(Params.with("countryCode", "US"));
 
             Assert.True(response != null);
             Assert.True(response.type == "covid19-area-report");
