+++ amadeus/dutyOfCare/diseases/TravelRestrictions.cs
